+++ src/main/java/com/unstoppabledomains/config/network/model/Contracts.java
@@ -1,0 +1,14 @@
+package com.unstoppabledomains.config.network.model;
+
+import java.util.Map;
+
+public class Contracts {
+    private Map<String, Contract> contracts;
+
+    private Contracts() {
+    }
+
+    public Map<String, Contract> getContracts() {
+        return contracts;
+    }
+}